[JP-357] Movendo pesquisa para outra pagina
